pyproject.toml
[project]
name = "ichatbio-nasa-power-agent"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "ichatbio-sdk==0.2.7",
    "openai~=1.86.0",
    "pydantic>=2.12.5",
    "instructor~=1.10.0",
    "dotenv~=0.9.9",
    "pytest>=8.3.5",
    "pytest-asyncio~=1.3.0",
    "pytest-httpx~=0.36.0",
    "uvicorn>=0.41",
    "zarr~=2.18.0",
    "s3fs~=2024.6.1",
    "numpy~=1.26.0",
    "xarray~=2024.9.0",
    "genson>=1.3.0",
    "geojson>=3.2.0",
    "jq>=1.11.0"
]

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
log_cli = true
log_cli_level = "INFO"